Added Spawn Points
Created two singletons: UIManager and SpawnPoint manager. The UIManager handles the button to claim the point to the user, and the SPManager handles the ownership of the spawnpoints for all users.

Created the SpawnPoint.cs script that handles information about the spawn point.

Current behaviour allows one to claim a spawnpoint and for it to begin producing (just debug statements) units via a coroutine.